# apps/wiki/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Plunk API",
    "description": "Open-source email platform API for transactional emails, campaigns, and marketing automation",
    "version": "1.0.0",
    "contact": {
      "name": "Plunk Support",
      "url": "https://www.useplunk.com"
    }
  },
  "servers": [
    {
      "url": "https://next-api.useplunk.com",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "API Key authentication. **Secret keys (sk_*)** are required for all endpoints except `/v1/track`. **Public keys (pk_*)** only work with the `/v1/track` endpoint for client-side event tracking. The project is automatically derived from the key."
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "time": {
            "type": "integer"
          }
        }
      }
    },
    "parameters": {
...
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 255
        },
        "description": "Optional key that guarantees this request runs at most once. If the key was already used by your project, the request is refused with `409` instead of being performed a second time. Keys are scoped to your project, expire after 24 hours (configurable when self-hosting), and must be 1-255 printable ASCII characters."
      }
    }
  },
  "paths": {
...
    "/v1/track": {
      "post": {
        "tags": ["Public API"],
        "summary": "Track event",
        "description": "Track an event for a contact. Automatically creates or upserts the contact, then records the event. Tracked events can be used as workflow triggers, segment filters, and audience filters.\n\n**Reserved event names** (rejected with `VALIDATION_ERROR` and code `reserved_event`): anything matching `email.*`, `contact.subscribed`, `contact.unsubscribed`, `segment.<slug>.entry`, `segment.<slug>.exit`. These are emitted by Plunk itself.\n\n**Idempotency**: re-tracking the same event creates a new event record. Send an `Idempotency-Key` header to have a repeated request refused with `409` instead.",
        "operationId": "trackEvent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "event"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Contact email. The contact is auto-created if it doesn't exist."
                  },
                  "event": {
                    "type": "string",
                    "description": "Event name. Cannot match the reserved patterns above."
                  },
                  "subscribed": {
                    "type": "boolean",
                    "description": "Subscription state to apply to the contact. **New** contacts default to subscribed (`true`). **Existing** contacts keep their current state unless you pass an explicit value here. Pass `false` to track an event without resubscribing an unsubscribed contact."
                  },
                  "data": {
                    "type": "object",
                    "additionalProperties": true,
                    "description": "Contact data and one-off event variables. Persistent values (primitives, plain objects) are saved on the contact and become available as template variables. Pass `{ value, persistent: false }` for one-shot variables that should not be stored on the contact (e.g. order IDs, transaction details). `null` deletes a field. Empty strings are ignored. Reserved keys are filtered out — see the contacts concept page."
                  }
                }
              },
              "example": {
                "email": "user@example.com",
                "event": "purchase",
                "data": {
                  "product": "Premium Plan",
                  "amount": 99
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event tracked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "contact": {
                          "type": "string",
                          "description": "Contact ID"
                        },
                        "event": {
                          "type": "string",
                          "description": "Event ID"
                        },
                        "timestamp": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Idempotency-Key already used. The request was refused, not performed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "IDEMPOTENCY_KEY_REUSED",
                    "message": "Idempotency-Key \"order-1234-receipt\" has already been used",
                    "statusCode": 409,
                    "requestId": "8f14e45f-ceea-467a-9575-1f0f38e0b1c2",
                    "details": {
                      "key": "order-1234-receipt",
                      "originalRequest": "POST /v1/send",
                      "originalRequestAt": "2025-01-15T10:30:00.000Z",
                      "originalStatusCode": 200
                    },
                    "suggestion": "This Idempotency-Key was already used, so the request was refused rather than performed twice. Generate a new key for a genuinely new request."
                  },
                  "timestamp": "2025-01-15T10:31:00.000Z"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ]
      }
    },
    "/v1/verify": {
      "post": {
        "tags": ["Public API"],
        "summary": "Verify email address",
        "description": "Verify an email address for validity, check if it's from a disposable domain or personal email provider, verify MX records, and detect potential typos with suggestions.",
        "operationId": "verifyEmail",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email address to verify"
                  }
                }
              },
              "examples": {
                "validEmail": {
                  "summary": "Valid email address",
                  "value": {
                    "email": "user@gmail.com"
                  }
                },
                "typoEmail": {
                  "summary": "Email with potential typo",
                  "value": {
                    "email": "user@gmial.com"
                  }
                },
                "disposableEmail": {
                  "summary": "Disposable email address",
                  "value": {
                    "email": "user@tempmail.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email verification completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Always true for successful requests"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "email": {
                          "type": "string",
                          "format": "email",
                          "description": "Email address that was verified"
                        },
                        "valid": {
                          "type": "boolean",
                          "description": "Whether the email appears to be valid overall"
                        },
                        "isDisposable": {
                          "type": "boolean",
                          "description": "Whether the email is from a disposable/temporary email domain"
                        },
                        "isAlias": {
                          "type": "boolean",
                          "description": "Whether the email is from a forwarding/alias service"
                        },
                        "isTypo": {
                          "type": "boolean",
                          "description": "Whether a potential typo was detected in the email address"
                        },
                        "isPlusAddressed": {
                          "type": "boolean",
                          "description": "Whether the email uses plus addressing (contains a + in the local part)"
                        },
                        "isPersonalEmail": {
                          "type": "boolean",
                          "description": "Whether the email is from a personal/free email provider (Gmail, Hotmail, Yahoo, etc.)"
                        },
                        "domainExists": {
                          "type": "boolean",
                          "description": "Whether the domain exists in DNS (has NS records)"
                        },
                        "hasWebsite": {
                          "type": "boolean",
                          "description": "Whether the domain has a website (has DNS A or AAAA records) - informational only"
                        },
                        "hasMxRecords": {
                          "type": "boolean",
                          "description": "Whether the domain has MX records configured for email delivery"
                        },
                        "suggestedEmail": {
                          "type": "string",
                          "format": "email",
                          "description": "Suggested correction if a typo was detected (optional)",
                          "nullable": true
                        },
                        "reasons": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "description": "Array of human-readable reasons describing the verification results"
                        }
                      },
                      "required": [
                        "email",
                        "valid",
                        "isDisposable",
                        "isAlias",
                        "isTypo",
                        "isPlusAddressed",
                        "isPersonalEmail",
                        "domainExists",
                        "hasWebsite",
                        "hasMxRecords",
                        "reasons"
                      ]
                    }
                  }
                },
                "examples": {
                  "validEmail": {
                    "summary": "Valid email",
                    "value": {
                      "success": true,
                      "data": {
                        "email": "user@gmail.com",
                        "valid": true,
                        "isDisposable": false,
                        "isAlias": false,
                        "isTypo": false,
                        "isPlusAddressed": false,
                        "isPersonalEmail": true,
                        "domainExists": true,
                        "hasWebsite": true,
                        "hasMxRecords": true,
                        "reasons": ["Email appears to be valid"]
                      }
                    }
                  },
                  "typoDetected": {
                    "summary": "Email with typo detected",
                    "value": {
                      "success": true,
                      "data": {
                        "email": "user@gmial.com",
                        "valid": false,
                        "isDisposable": false,
                        "isAlias": false,
                        "isTypo": true,
                        "isPlusAddressed": false,
                        "isPersonalEmail": false,
                        "domainExists": false,
                        "hasWebsite": false,
                        "hasMxRecords": false,
                        "suggestedEmail": "user@gmail.com",
                        "reasons": [
                          "Possible typo detected, did you mean gmail.com?",
                          "Domain does not exist (no nameservers found)"
                        ]
                      }
                    }
                  },
                  "disposableEmail": {
                    "summary": "Disposable email detected",
                    "value": {
                      "success": true,
                      "data": {
                        "email": "user@tempmail.com",
                        "valid": true,
                        "isDisposable": true,
                        "isAlias": false,
                        "isTypo": false,
                        "isPlusAddressed": false,
                        "isPersonalEmail": false,
                        "domainExists": true,
                        "hasWebsite": true,
                        "hasMxRecords": true,
                        "reasons": ["Email appears to be valid"]
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid email format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}